requirements.txt
# Golf Intelligence Dashboard Dependencies
# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0
altair>=5.0.0

# Data Processing
pyarrow>=14.0.0
duckdb>=0.9.0

# Utilities
pyyaml>=6.0.0
python-dateutil>=2.8.0

# Statistical Analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Google Sheets connectivity
# gspread>=5.10.0
# oauth2client>=4.1.3
